fix linter

+++ deployment/ccip/changeset/testhelpers/test_sui_helpers.go
@@ -731,7 +731,7 @@ func UpgradeContractDirect(
 
 	paramValues := []any{
 		suiBind.Object{Id: upgradeCapID},
-		uint8(policy),
+		policy,
 		digest,
 	}
